chore: move search-packages files
