+++ lib/src/modules/search_packages/data/models/events_model.dart
