+++ lib/src/modules/search_packages/data/models/packages_model.dart
