+++ lib/src/modules/search_packages/presentation/controller/states/search_packages_states.dart
@@ -1,4 +1,4 @@
-import '../../data/models/packages_model.dart';
+import '../../../data/models/packages_model.dart';
 
 abstract interface class SearchPackagesState {}
