+++ lib/src/modules/search_packages/presentation/widgets/search_package_header.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 
-import '../../../../core/controller/search_packages_controller.dart';
+import '../controller/search_packages_controller.dart';
 import '../../../../core/themes/app_colors.dart';
 
 class SearchPackageHeaderWidget extends StatelessWidget {
