+++ test/src/core/data/search_packages_repository_test.dart
@@ -3,8 +3,8 @@ import 'dart:convert';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:result_dart/result_dart.dart';
-import 'package:your_tracker/src/core/data/models/packages_model.dart';
-import 'package:your_tracker/src/core/data/repositories/search_packages/search_packages_repository.dart';
+import 'package:your_tracker/src/modules/search_packages/data/models/packages_model.dart';
+import 'package:your_tracker/src/modules/search_packages/data/repositories/search_packages/search_packages_repository.dart';
 import 'package:your_tracker/src/core/services/http/http.dart';
 import 'package:your_tracker/src/core/services/http/uno.dart';
 import 'package:your_tracker/src/core/services/local_storage/flutter_secure_storage.dart';
